fix(git): move branch_exists regression to own integration test binary
Review round 4 finding from CodeRabbit on 33fa8028:

The regression test at src/git/worktree.rs used `#[serial(path_env)]`
+ process-global `PATH=""` to force a spawn failure. That worked
against tests sharing the `path_env` slot but did not serialize
against the ~20+ other git-spawning tests in the same lib binary
which either have no `#[serial]` or use a different slot key
(`#[serial(submodule)]`). Under `cargo test`'s default parallel
runner, `PATH=""` in one thread could make concurrent
`Command::new("git")` calls in sibling tests fail with ENOENT
non-deterministically. Latent CI flake.

Options considered:

- Annotate every git-spawning test in the file with
  `#[serial(path_env)]`: cross-cutting change touching ~20+ tests
  for a single regression, and no compile-time enforcement against
  a future test being added without the slot.
- Fake `git` shim: requires refactoring `run_git` to accept a
  custom Command env, invasive for a test-only concern.
- Move the test to its own integration test binary: each
  `tests/*.rs` file at the top level of `tests/` gets its own
  process, so process-global env mutation is isolated by
  construction. No cross-file races possible. Matches the pattern
  of `tests/plugin_install.rs` and siblings that mutate HOME /
  XDG_CONFIG_HOME. Zero cross-cutting annotation drift.

Picked option 3. `tests/branch_exists_spawn_failure.rs` runs alone
in its own binary; `#[serial(path_env)]` inside that binary is
retained purely as intra-file future-proofing for additional
PATH-mutating tests. The test body is otherwise unchanged
(`PathGuard` RAII, variant-only assertions, cross-refs to the
integration tests locking `Ok(true)`/`Ok(false)`).

Refs #2653.
